chore: refactor type to `types.py`
